requirements.txt
PyAutoGUI==0.9.50
pyserial==3.4